# docs/swagger.yaml
definitions:
  handler.AmenityCreate:
    properties:
      name:
        type: string
    type: object
  handler.AmenityUpdate:
    properties:
      name:
        type: string
    type: object
...
  handler.ErrorBadRequest:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  handler.ErrorInternal:
    properties:
      error:
        example: some internal error
        type: string
    type: object
  handler.ErrorNotFound:
    properties:
      error:
        example: not found
        type: string
    type: object
  handler.FavoriteCreate:
    properties:
      listing_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  handler.StatusOK:
    properties:
      message:
        example: deleted successfully
        type: string
    type: object
...
  model.Amenity:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.Favorite:
    properties:
      id:
        type: integer
      listing_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  model.ListingAmenity:
    properties:
      amenity_id:
        type: integer
      listing_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: DB CW
  version: "1.0"
paths:
  /api/amenities:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Amenity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить все удобства
      tags:
      - amenities
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные удобства
        in: body
        name: amenity
        required: true
        schema:
          $ref: '#/definitions/handler.AmenityCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Amenity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Создать удобство
      tags:
      - amenities
  /api/amenities/{id}:
    delete:
      parameters:
      - description: Amenity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Удалить удобство
      tags:
      - amenities
    get:
      parameters:
      - description: Amenity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Amenity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить удобство по ID
      tags:
      - amenities
    put:
      consumes:
      - application/json
      parameters:
      - description: Amenity ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные удобства
        in: body
        name: amenity
        required: true
        schema:
          $ref: '#/definitions/handler.AmenityUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Amenity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Обновить удобство
      tags:
      - amenities
...
  /api/favorites:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные избранного
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/handler.FavoriteCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Favorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Добавить объявление в избранное
      tags:
      - favorites
  /api/favorites/{id}:
    delete:
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Удалить избранное по ID
      tags:
      - favorites
    get:
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Favorite'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить избранное по ID
      tags:
      - favorites
...
  /api/listings/{listing_id}/amenities:
    get:
      parameters:
      - description: Listing ID
        in: path
        name: listing_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Amenity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить удобства объявления
      tags:
      - amenities
  /api/listings/{listing_id}/amenities/{amenity_id}:
    delete:
      parameters:
      - description: Listing ID
        in: path
        name: listing_id
        required: true
        type: integer
      - description: Amenity ID
        in: path
        name: amenity_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Удалить удобство из объявления
      tags:
      - amenities
    post:
      consumes:
      - application/json
      parameters:
      - description: Listing ID
        in: path
        name: listing_id
        required: true
        type: integer
      - description: Amenity ID
        in: path
        name: amenity_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ListingAmenity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Добавить удобство к объявлению
      tags:
      - amenities
...
  /api/users/{user_id}/favorites:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Favorite'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Получить избранное пользователя
      tags:
      - favorites
  /api/users/{user_id}/favorites/{listing_id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Listing ID
        in: path
        name: listing_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusOK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorInternal'
      summary: Удалить избранное по user_id и listing_id
      tags:
      - favorites
...
swagger: "2.0"